# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Allmax Backend API",
    "version": "1.0.0",
    "description": "API documentation for Allmax Backend"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/shop/goods/list-with-price": {
      "post": {
        "summary": "Get goods list with price",
        "tags": [
          "Shop"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pageSize": {
                    "type": "integer",
                    "default": 20
                  },
                  "pageNumber": {
                    "type": "integer",
                    "default": 1
                  },
                  "search": {
                    "type": "string",
                    "description": "Search keyword for product name"
                  },
                  "cateId": {
                    "type": "string",
                    "description": "Category ID to filter by"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "records": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          },
          "503": {
            "description": "Database not connected"
          }
        }
      }
    },
    "/shop/getAllSkuWithAttrValues": {
      "post": {
        "summary": "Get all SKUs with attribute values for a SPU (POST)",
        "tags": [
          "Shop"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "spuId"
                ],
                "properties": {
                  "spuId": {
                    "type": "string",
                    "description": "SPU ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SKUs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing spuId"
          },
          "500": {
            "description": "Server error"
          },
          "503": {
            "description": "Database not connected"
          }
        }
      },
      "get": {
        "summary": "Get all SKUs with attribute values for a SPU (GET)",
        "tags": [
          "Shop"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "spuId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "SPU ID"
          }
        ],
        "responses": {
          "200": {
            "description": "SKUs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing spuId"
          },
          "500": {
            "description": "Server error"
          },
          "503": {
            "description": "Database not connected"
          }
        }
      }
    },
...
  },
...
}